move_or_reorder_blocks_hit_testing_dump_records
